backend/requirements.txt
fastapi==0.115.0 
uvicorn==0.30.6 
pydantic==2.9.2 
sqlalchemy==2.0.35 
psycopg2-binary==2.9.9 
pyjwt==2.9.0 
requests==2.32.3 
moviepy==1.0.3 
python-jose==3.3.0 
passlib[bcrypt]==1.7.4